fix: correct grammar and simplify container structure
