+++ src/design/modals/detailsRequestsInProgressModal/styles.ts
@@ -29,14 +29,6 @@ const StyledModalConatiner = styled.div`
   padding: ${basic.spacing.s200};
 `;
 
-const StyledContainerButton = styled.div`
-  & button {
-    display: flex;
-    padding-right: ${basic.spacing.s0};
-    justify-content: flex-end;
-  }
-`;
-
 const StyledContainerData = styled.div`
   display: flex;
   flex-direction: column;
@@ -45,13 +37,7 @@ const StyledContainerData = styled.div`
   padding: ${basic.spacing.s075} ${basic.spacing.s150} ${basic.spacing.s075}
     ${basic.spacing.s150};
   border-radius: 8px;
-  border: 1px solid ${inube.palette.neutral.N30};
   background-color: ${inube.palette.neutral.N10};
 `;
 
-export {
-  StyledModal,
-  StyledContainerButton,
-  StyledModalConatiner,
-  StyledContainerData,
-};
+export { StyledModal, StyledModalConatiner, StyledContainerData };
