+++ src/types/positions/details/IDetailsModal/index.ts
@@ -1,0 +1,9 @@
+import { IPosition } from "@ptypes/positions/assisted/IPosition";
+import { IField } from "../IField";
+
+interface IDetailsModal {
+  data?: IPosition;
+  labelsOptions: IField[];
+}
+
+export type { IDetailsModal };
