+++ src/types/positions/details/IField/index.ts
@@ -1,0 +1,5 @@
+interface IField {
+  id: string;
+  labelName: string;
+}
+export type { IField };
